[runtime][hip] Fix format errors and conflicting types.
The `iree_make_status` requires the third argument to be `unsigned long`
type, while the `iree_hal_queue_affinity_t` is a `unsigned long long`
type. The revision casts it to `unsigned long` type to fix the error
from `-Wformat`.

```
runtime/src/iree/hal/drivers/hip/hip_multi_queue_command_buffer.c:120:29:
error: format specifies type 'unsigned long' but the argument has type
'iree_hal_queue_affinity_t' (aka 'unsigned long long') [-Werror,-Wformat]
```

The revision also fixes the conflicting types for
`iree_hal_hip_stream_command_buffer_create`. The order was wrong.

+++ runtime/src/iree/hal/drivers/hip/hip_multi_queue_command_buffer.c
@@ -117,7 +117,7 @@ IREE_API_EXPORT iree_status_t iree_hal_hip_multi_queue_command_buffer_get(
   if (!(command_buffer->base.queue_affinity & queue_affinity)) {
     return iree_make_status(IREE_STATUS_NOT_FOUND,
                             "no command buffer for affinity %lu",
-                            queue_affinity);
+                            (unsigned long)queue_affinity);
   }
   int index = iree_math_count_ones_u64(command_buffer->base.queue_affinity &
                                        (queue_affinity - 1));
